requirements.txt
# Core Framework
langchain
langchain-community
langchain-openai

# YouTube Data Extraction
youtube-transcript-api
pytube
langchain
# ... (rest of your previous requirements)
# Vector Database (Using ChromaDB)
chromadb

# Utilities
python-dotenv
tqdm
tiktoken